// api/model/createDetailDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * POCHY API 명세서
 * POCHY Swagger UI
 * OpenAPI spec version: 1.0.0
 */

export interface CreateDetailDto {
  name?: string;
  brand?: string;
  category?: string;
  subCategory?: string;
  price?: number;
  feature?: string;
  memo?: string;
  productImageUrl?: string;
}